fix: optimal↔dispatch end-to-end consistency (Codex gpt-5.5 comprehensive review, 7 findings)
Applied by Codex (tmux) from its review; reviewed + verified here. Makes the optimal
table actually match what runtime dispatch can select:
1. optimal_order(): no provider outside a non-empty table chain sorts before chain[0]
   (kernel-set-only chain => kernel-set first) — measured kernel-set winners now win.
2/3. gen_optimal.validate() is a STRONG gate: every non-kernel-set winner + chain entry
   must be CALLABLE (call != None), sm>=min_sm, and pass hw-dtype + provider.dtypes;
   fixed the 17 non-callable winners (impl flashinfer gemma_rmsnorm/sampling adapters
   or de-rank) and the min_sm/dtype-gated winners (torch-sdpa@sm75, flash-attn@fp8,
   sgl-mla@fp8, vllm-moe@int4).
4. models/select.py optimal lookup keyed by (plan_op, resolved scheme): fp8->fp8_gemm,
   w8a8->int8_gemm, w4a16->w4a16 (no more dense-GEMM mis-annotation; ksctl plan fp8 now
   shows DeepGEMM fp8).
5. registry-import failure is fatal in source-tree gen; cells validated vs real Provider
   objects (not just names).
6. one Blackwell contract: sm120->sm100 capability family folded in a shared helper used
   by select_optimal + dispatch; rtx5090 mapping consistent.
7. optimal.py loads sm_thresholds from gpu_caps.json (single source).
Verified: gen_optimal --check (strong) clean; 70 dispatch+optimal tests pass; ksctl
plan/optimal/backends consistent.

+++ bindings/python/tests/test_dispatch.py
@@ -232,16 +232,16 @@ def test_sgl_kernel_wins_moe_gate_ops_when_available(monkeypatch):
 def test_deepgemm_wins_moe_grouped_gemm_on_hopper(monkeypatch):
     # On sm90 DeepGEMM grouped FP8 is the rank-1 MoE GEMM (DeepSeek-V3 path).
     _mock_available(monkeypatch, available_libs={"deep_gemm"}, sm=90)
-    assert dispatch.which("moe") == "deep_gemm"
+    assert dispatch.which("moe", dtype="fp8") == "deep_gemm"
     # sgl-kernel grouped GEMM is the rank-2 alignment target on sm90.
     dispatch.reset_cache()
     _mock_available(monkeypatch, available_libs={"sgl_kernel"}, sm=90)
-    assert dispatch.which("moe") == SGL_KERNEL
+    assert dispatch.which("moe", dtype="fp8") == SGL_KERNEL
     # DeepGEMM + sgl + vllm all present on sm90 -> DeepGEMM (rank 1) wins.
     dispatch.reset_cache()
     _mock_available(
         monkeypatch, available_libs={"deep_gemm", "sgl_kernel", "vllm"}, sm=90)
-    assert dispatch.which("moe") == "deep_gemm"
+    assert dispatch.which("moe", dtype="fp8") == "deep_gemm"
 
 
 def test_vllm_fused_moe_on_ampere(monkeypatch):
@@ -366,7 +366,7 @@ def test_compute_bound_prefers_external_when_available(monkeypatch):
         ("w4a16", {"vllm"}, 80, "int4", "vllm-marlin"),
         ("w4a16", {"vllm"}, 90, "int4", "vllm-machete"),
         ("attention_prefill", {"flash_attn"}, 80, "bf16", "flash-attn"),
-        ("attention_decode", {"flashinfer"}, 80, "bf16", "flashinfer"),
+        ("attention_decode", {"flashinfer"}, 80, "fp16", "flashinfer"),
         ("mla_decode", {"sgl_kernel"}, 90, "bf16", SGL_KERNEL),
         ("mla_decode", {"flashinfer"}, 80, "bf16", "flashinfer"),
         ("moe", {"deep_gemm"}, 90, None, "deep_gemm"),
